requirements.txt
pytest
pydantic
openai
python-dotenv
sqlmodel
fastapi
uvicorn
jinja2
